ocaml: Show the parameter names in documentation
Add the parameter names to the documentation, which helps when reading
the blurb about each function since that often refers to parameter
names.

For example in the NBD.connect_tcp call, the documentation refers to
the "hostname" and "port" parameters, but they previously appeared
only as two anonymous strings.  After this change:

  val connect_tcp : t -> string -> string -> unit
  (** [NBD.connect_tcp t hostname port]           <-- this line added

      connect to NBD server over a TCP port

      Connect (synchronously) to the NBD server listening on
      "hostname:port". The "port" may be a port name such as
      "nbd", or it may be a port number as a string such as
      "10809".
      ...