ocaml: More accurate errno representation in exceptions
OCaml has a helper function for converting C errno to OCaml Unix.error
types (in fact it was added in 2005).  Use this function to convert
the raw C errno from NBD errors into a natural type in exceptions.

Since nbd_get_errno doesn't always have data to return, the final type
is “Unix.error option”.

This is an API change for OCaml callers, but we don't guarantee the
API stability for non-C languages.