ocaml: Allow OCaml callbacks to update !err with C errno
The previous commit was not quite complete, because the tests relied
on errno 100 == ENETDOWN and 101 == ENETUNREACH.  This is probably
only true for Linux.

Ideally the callback parameter !err (containing an errno) would be
converted to and from a “Unix.error option” transparently.  However
the generator does not currently have sufficient meta-information for
this, and it's also rather complex to implement with large overhead
even if we did.

Instead allow callbacks to update !err with the C errno (or 0 for no
error).  To do this we simply need to allow access to the OCaml C
function code_of_unix_error, added in OCaml 4.01.0.

(cherry picked from commit 806a63ad82956b02dcaf540bc8c9f371f659922d)