tests: Test nbd_opt_list_meta_context() without SR
This test proves that an attempt to list meta contexts goes to the
server rather than being rejected at the client, even with older
nbdkit that fails the attempt[1]; and includes ports to the various
languages.  Done as a separate commit to allow temporary reordering of
the series to prove the test works.

[1] nbdkit didn't accept LIST_META_CONTENT before STRUCTURED_REPLY
until commit c39cba73 [v1.27.3, backported to 1.26.6].  Similarly,
qemu-nbd didn't accept it until commit da24597d [v6.2].  So, I was
able to test the graceful skip by adding a build of nbdkit without
that support at the front of my PATH.

Acked-by: Richard W.M. Jones <rjones@redhat.com>