debug: Trace even may_set_error=false functions
When we introduced debug tracing, we stated that "because of
limitations in the way debugging works it can only be done on
functions which have may_set_error = true".  This is because our trace
statements depended on the current error context for part of its
trace.

But this is overly restrictive; if we know there is not going to be
any error, we can just pass in the function name to the debug trace
directly, by tweaking the signature for internal debugging.  The
common case of debug() statements depending on an implicit context is
unchanged.

An example of the changes to the generated lib/api.c is:

| --- lib/api.c.bak       2022-09-03 11:07:52.668893312 -0500
| +++ lib/api.c   2022-09-03 11:19:19.793796181 -0500
| @@ -68,8 +68,16 @@ nbd_get_debug (struct nbd_handle *h)
|
|    /* This function must not call set_error. */
|    pthread_mutex_lock (&h->lock);
| +  if_debug (h) {
| +    debug_direct (h, "nbd_get_debug", "enter:");
| +  }
| +
|    ret = nbd_unlocked_get_debug (h);
|
| +  if_debug (h) {
| +    debug_direct (h, "nbd_get_debug", "leave: ret=" "%d", ret);
| +  }
| +
|    if (h->public_state != get_next_state (h))
|      h->public_state = get_next_state (h);
|    pthread_mutex_unlock (&h->lock);

and traces for functions like set_private_data() now show up in the
logs.

Fixes: 9df088bf ("lib: Print function call parameters and results in debug output.", v0.9.7)
(cherry picked from commit 1f16fcf03249e3832f72971ab735b49570de69bd)