api: Localize list used during NBD_OPT_*_META_CONTEXT
Right now, our generator ensures that we are in the Negotiating state
before a client can call nbd_add_meta_context() while there is an
active server connection.  That is, even if the user piles up enough
meta contexts that a slow-response server then causes nbd_aio_opt_go()
and friends to return early on blocked writes, with the state machine
now tracking pointers into the middle of the global list, those
pointers are not at risk of becoming a use-after-free because the user
cannot successfully call nbd_add_meta_context() to alter the global
list until they first resume the state machine (using nbd_aio_poll or
similar) to finish flushing the client's write to the server and
eventually get back to the Negotiating state.

However, it took me a while to perform that audit to ensure that the
user can't mess with the global list while the state machine is
peering into the middle of it (that is, state machine interlocks act
at a distance).  Easier is to have the state machine act on a local
copy, which cannot be directly altered by user action regardless of
the current state; and therefore even if we later change our mind
about which states can call nbd_add_meta_context() (the state
restrictions in API.ml), that change won't accidentally cause our
state machine to start seeing corrupted list contents on the uncommon
case of blocked writes (in states-newstyle-opt-meta-context.c).

In this patch, I added a helper function to copy from either a
StringList (no clients of that parameter in this patch, but coming
soon) or from the implicit list.  Also, at the time of this patch, the
copying is done once in the state machine, but later patches will
hoist it earlier into lib/opt.c as part of adding more APIs for
fine-grained control over nbd_set_opt_mode management of context
lists, starting with nbd_opt_list_meta_context_queries() which will
take an explicit list of meta contexts to query.

Acked-by: Richard W.M. Jones <rjones@redhat.com>