api: Add nbd_[aio_]opt_list_meta_context_queries
Rather than always relying on the implicit global list managed by
nbd_add_meta_context() and nbd_clear_meta_contexts(), we can take the
user's list of contexts as a direct parameter.  This finally makes use
of the non-NULL queries parameter of the nbd_internal_set_querylist()
added in a previous patch.

The C test is added along with this commit (applying the new test
independently wouldn't compile), but the language bindings are done
separately for ease of review.

For now, passing NULL instead of an empty list is not documented
behavior and should not be relied on.  The python bindings don't
support it (trying to pass None instead of [] results in a TypeError),
and in C code, it happens to reuse the global nbd_add_meta_context
list (but we may change it later to fail with EFAULT, or to trigger an
assertion failure, or even to treat a NULL list the same as an
explicit list with 0 elements; discussion is still ongoing on the
list).

Acked-by: Richard W.M. Jones <rjones@redhat.com>