api: Add nbd_[aio_]opt_set_meta_context[_queries]
Make it possible for the client to choose when to manually request
meta contexts when using nbd_set_opt_mode().  By itself, this patch
allows us to test how a server reacts to NBD_OPT_SET_META_CONTEXT
without a previous negotiation of structured headers; then an upcoming
patch will also add APIs for choosing when to negotiate OPT_STARTTLS
and OPT_STRUCTURED_REPLY for even more fine-grained testing control.
Most users won't need this API (the automatic defaults are good
enough), but when writing a server, this level of integration testing
can be useful.  Similar to the existing nbd_opt_list_meta_context
API, also offer a variant that takes an explicit query list.

As this is a new API, the unit test in C is included in this patch;
there is no point in splitting it as reordering the patches would not
compile.  However, porting the tests to other languages will be done
in the next commit.

As with the recently-added nbd_opt_list_meta_context_queries, the
behavior of a NULL list argument is undocumented, and may change once
list discussion settles on what that behavior should be.

Acked-by: Richard W.M. Jones <rjones@redhat.com>