api: Add nbd_[aio_]opt_structured_reply()
Recent commits have lamented that we do not have a way to provoke
atypical ordering of option negotiation during server integration
testing.  This commit adds nbd_opt_structured_reply() to make it
possible to fine-tune when structured replies are requested, and fills
in a couple of FIXME comments left in previous commits.  The new API
can be called more than once; the testsuite is also enhanced with a
test that shows that nbdkit rejects the second attempt but libnbd
still remembers the first one succeeded (this requires moving the wipe
of h->structured_reply out of NBD_OPT_STRUCTURED_REPLY error handling,
and instead putting it into NBD_OPT_STARTTLS success, which becomes
important when a later patch adds nbd_opt_starttls()).

The C test changes depend on a new API, so they can't reasonably be
applied out of order.  However, porting the FIXME changes of the
testsuite to other language bindings will be done in the next patch,
to ease review of this one.  The test opt-structured-twice borders on
being an integration test (in fact, nbdkit commit 9418ec58 [upcoming
v1.34] changes nbdkit from accepting a second structured reply request
to instead diagnosing it as redundant; both behaviors are compatible
with the NBD spec), so that one will remain C only.

Acked-by: Richard W.M. Jones <rjones@redhat.com>