crypto: Centralize TLS debug info
Commit 2d4cbfb1 added some additional debug data when the TLS
handshake concludes, except it only does so if the last part of the
handshake was a read.  If (for whatever reason) handshake ends with a
write, we lose out on the debug details; given that no one has
complained, it looks like this doesn't usually happen.  Still, it is
just as easy to create a new state to consolidate both conclusion
paths through the same debug output.

Fixes: 2d4cbfb1 ("states: When the connectoin has been upgraded to TLS, print the cipher etc.", v0.1)
(cherry picked from commit 8685ab673b271fe6ce335b8c23470b32e0d7dfc0)