lib: Fix memory leak on (unlikely) socket activation failure
Compiling with -fanalyzer was noisy with false positives, but among
the noise, it spotted a real bug:

generator/states-connect-socket-activation.c:87:6: warning: leak of ‘p’ [CWE-401] [-Wanalyzer-malloc-leak]
  ‘prepare_socket_activation_environment’: event 1
... <lots more, and hard to read because the compiler gets lost in the
combination of mapping the generated .c back to our #line-mapped input
.c file and macro expansions of vector.h implementation>

Looking deeper, when we refactored the code to use vectors, we lost
the code that cleans up partial allocations if the overall
prepare_socket_activation_environment() call fails.  Note, however,
that failing to clean up memory after failing to append to a list is
unlikely in all but the worst situations where we are likely to remain
OOM anyways.

Fixes: 6f3eee2e ("lib: Replace a few uses of realloc with nbdkit vector library.", v1.5.5)