lib: Avoid leak if nbd_internal_set_argv is used twice
Commit 21b279a78 plugged a leak found by valgrind when copying a list
more than once, easily possible with the recent addition of
nbd_opt_list_meta_context_queries().  Harder is coming up with a
scenario where nbd_connect_command() can be called more than once on
the same handle (on success, we move the state machine to
Negotiating/Ready; on failure, we move the state machine to Dead;
either way, we can't attempt to connect to another command).  In spite
of that, gcc -fanalyzer was smart enough to point out that since a
string_vector contains potential pointers in the vector, our use of
string_vector_reset() can potentially leak those pointers left over
from an earlier call to nbd_internal_copy_string_list().  For
consistency with the previous leak fix in utils.c, change the contract
slightly: now all callers about to copy a string vector ensure that
the destination is empty on entry.