ci: Add coverage of --without-libxml2
Although we generally like URI support, it is worth having explicit
coverage in our CI that we can build without it for those who want
minimal dependencies.  Similar to the previous patch, this is easiest
to do with another build.sh variable.  Doing so found two additional
tests that need skipping (comparable to how we already skipped similar
tests in sh/).