ocaml: Prefer caml_unix_get_sockaddr over get_sockaddr
OCaml 5.00 fixes a problem mentioned in commit 50500aade9, that the
get_sockaddr symbol is not namespaced.  Use a configure-time test to
see if the new namespaced symbol is available, but fall back to the
old one if not.

(cherry picked from commit be92af8f533b5625fadae4c910f7bb15a6fd59f3)