ocaml: Simplify the sockaddr test
The sockaddr test added in commit 50500aade9 ("ocaml: Implement
sockaddr type") was quite complicated because it had to run nbdkit as
a subprocess, manage the socket, check for a PID file etc.

The simpler way of doing this is to make it work like the Python test
(python/test-aio-connect-unix.sh) where we run
nbdkit -U - ... --run './test $unixsocket'

The straightforward way would be to a separate shell script etc to
ocaml/tests/Makefile.am:TESTS which is orthogonal to how the existing
tests work.  So instead make the test exec nbdkit when called first
time (without the $unixsocket parameter) and then run the test when
called the second time.

Updates: commit 50500aade9f32899eddd2a7b89ae5c596674f95c
Acked-by: Eric Blake <eblake@redhat.com>