api: Add LIBNBD_STRICT_PAYLOAD to honor qemu 32M write limit
Modern qemu tends to advertise a strict 32M payload limit.  The NBD
spec says that if the user did not request strict sizes, the server's
advertisement is merely advisory, and that meanwhile portable clients
should not try larger than 32M.  But it also states that when the
client DOES request strict sizes (which is libnbd's default since
commit 0a9adf6b in v1.12.0), then attempts to request oversize reads
or writes trigger undefined behavior.  Our design goal for libnbd has
been that we should avoid undefined behavior by default, even if we
allow user opt-in to override our safety checks for the purposes of
interoperability testing of server features.

Yet we were unconditionally defaulting to allowing the user to attempt
up to 64M in pwrite.  For pread, qemu will reply with EINVAL but keep
the connection up; but for pwrite, an overlarge buffer is fatal.
Killing the connection because we triggered undefined behavior from
the server is not nice.  It's time to teach libnbd to honor qemu's max
buffer by default, while still allowing the user to disable the strict
mode setting if they want to try 64M anyways.  However, the XXX
comment in internal.h about letting the user configure a larger limit
than 64M remains a task for another day.

This also involves advertising a new LIBNBD_SIZE_PAYLOAD via
nbd_get_block_size, which is more reliable than asking the user to
manually obey LIBNBD_SIZE_MAXIMUM which may not be set or may be
larger than 64M.

Message-Id: <20221109232607.1390404-2-eblake@redhat.com>
[commit message improved, use PRIu32 instead of %d, fix documented constant]
Reviewed-by: Laszlo Ersek <lersek@redhat.com>