Revert "lib: account for realpath deficiency on some platforms"
This reverts commit b29ff42e5d005abdfd837d7a8b5fa9f0720ad469.

Turns out the limitation that "realpath" does not recognize the "--"
end-of-options argument separator is just one kind of special sauce that
Alpine Linux / BusyBox gift to the world. I'm going to disable
"lib/test-fork-safe-execvpe.sh" on Alpine Linux altogether (Alpine Linux
is unsalvageable) in a subsequent patch; for now, restore the "realpath"
invocation to its right form.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>