lib/test-fork-safe-execvpe.sh: generalize "run" to "run0"
It turns out that we'll need to put the generality of our
"test-fork-safe-execvpe" binary executable to use, in that it takes
separate "program-to-exec" and "argv0" (for the program-to-exec)
arguments.

Currently, the "run" function duplicates $2 to both "program-to-exec" and
"argv0", for "test-fork-safe-execvpe". Remove the duplication (expect the
caller to provide separate $2 and $3 arguments, respectively) and rename
"run" to "run0". At the same time, reimplement "run" as a simple wrapper
(i.e., with just the duplication) around "run0".

This patch is worth viewing with "git show --color-words" as well.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>