lib/test-fork-safe-execvpe.sh: cope with Alpine Linux / BusyBox limitations
Whenever we expect our "bin/f" and "sh/f" files to execute (and to
succeed), provide "expr" and "sh" as argv[0], respectively. This is to
work around BusyBox, where the central busybox binary selects its behavior
based on argv[0]. See e.g.
<https://bugs.busybox.net/show_bug.cgi?id=15481>.

Also cope with "realpath" (from BusyBox) not accepting the "--"
end-of-options argument separator. See
<https://bugs.busybox.net/show_bug.cgi?id=15466>.

This patch is worth viewing with "git show --color-words" as well.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>